First Aider Coding Assignment
This was my first time using Aider, so I learned a lot of its basic abilities. Aider conducted a step-by-step tutorial for me to modify an element in a block. I first defined the variable "json_string" as a string reference to a dictionary (i.e. a JSON string), and then I told Aider that I wanted to modify the block. Aider told me to assign the variable "blocks" with a dictionary version of "json_string" with the command "blocks = json.loads(json_string)". Then, I typed the command "blocks["size"] = 8000" to replace the original size 9195 with 8000. Lastly, I typed "json_string = json.dumps(blocks)" to update "json_string" from its original version (size=9195) to its new version (size=8000). 

I have also learned from Aider how to plot a histogram of transactions, as shown in the code here. I told Aider that I wanted to plot a histogram of transactions, so Aider asked if I would like it to write a new Python coding file to plot the histogram. I answered yes, so it created a new file named "transaction_histogram.py". I copied this file and pasted it to the code shown here. 

I have used Github very little before this class, so I feel new to it. Therefore, I spent plenty of time attempting to figure out how to sync Aider with Github, so I might not have satisfied all conditions for this assignment. Hopefully by the end of this semester, I will be fluent with Github, which will be important to my future computer science classes and possibly my future career projects.