aider: Replace "#" with "javascript:void(0)" in Navigation component.
# Context:
USER: I'm geting this as an error """  Line 10:7:  The href attribute requires a valid value to be accessible. Provide a valid, navigable address as the href value. If you cannot provide a valid href, but still need the element to resemble a link, use a button and change it with appropriate styles. Learn more: https://github.com/jsx-eslint/eslint-plugin-jsx-a11y/blob/HEAD/docs/rules/anchor-is-valid.md  jsx-a11y/anchor-is-valid"""
ASSISTANT: The error message is indicating that the anchor tags in your Navigation component do not have valid href values. The href attribute in an anchor tag is used to specify the URL of the page the link goes to. In your case, you're using "#", which is not a valid URL.

To fix this, you can replace the "#" with a valid URL or a path to another page in your application. If you don't have a specific URL to use, you can use "javascript:void(0)" as a placeholder. This will prevent the page from refreshing when the link is clicked, which is the default behavior when "#" is used.

Here's how you can modify your code:

```tsx
workspaces/mySwap/src/Navigation.tsx
<<<<<<< ORIGINAL
    <nav className="uniswap-navigation">
      <a href="#">Swap</a>
      <a href="#">Pool</a>
      <a href="#">Analytics</a>
      <a href="#">Docs</a>
    </nav>
=======
    <nav className="uniswap-navigation">
      <a href="javascript:void(0)">Swap</a>
      <a href="javascript:void(0)">Pool</a>
      <a href="javascript:void(0)">Analytics</a>
      <a href="javascript:void(0)">Docs</a>
    </nav>
>>>>>>> UPDATED
```

Please replace "javascript:void(0)" with the actual URLs when you have them.