aider: Add the crypto library to the package.json.
# Context:
USER: add the crypto library to the package.json
ASSISTANT: To add the `crypto` library to the `package.json` file, we will modify the `workspaces/cross-chain_composition/package.json` file.

Here are the changes:

```json
workspaces/cross-chain_composition/package.json
<<<<<<< ORIGINAL
{
    "name": "cross-chain_composition",
    "version": "1.0.0",
    "description": "",
    "main": "index.js",
    "scripts": {
        "test": "echo \"Error: no test specified\" && exit 1"
    },
    "author": "",
    "license": "ISC",
    "dependencies": {
    }
}
=======
{
    "name": "cross-chain_composition",
    "version": "1.0.0",
    "description": "",
    "main": "index.js",
    "scripts": {
        "test": "echo \"Error: no test specified\" && exit 1"
    },
    "author": "",
    "license": "ISC",
    "dependencies": {
        "crypto": "^1.0.1"
    }
}
>>>>>>> UPDATED
```

In this updated code, the `crypto` library is added to the `dependencies` section of the `package.json` file. The version of the `crypto` library is set to `^1.0.1`, which means any version that is `1.0.1` or higher, but less than `2.0.0`, is acceptable.
USER: add the crypto library to the package.json
ASSISTANT: To add the `crypto` library to the `package.json` file, we will modify the `workspaces/cross-chain_composition/package.json` file.

Here are the changes:

```json
workspaces/cross-chain_composition/package.json
<<<<<<< ORIGINAL
{
  "name": "cross-chain_composition",
  "version": "0.0.1",
  "description": "CPSC 370 assignments",
  "main": "index.ts",
  "scripts": {
    "test": "echo testing"
  },
  "author": "",
  "license": "ISC"
}
=======
{
  "name": "cross-chain_composition",
  "version": "0.0.1",
  "description": "CPSC 370 assignments",
  "main": "index.ts",
  "scripts": {
    "test": "echo testing"
  },
  "author": "",
  "license": "ISC",
  "dependencies": {
    "crypto": "^1.0.1"
  }
}
>>>>>>> UPDATED
```

In this updated code, the `crypto` library is added to the `dependencies` section of the `package.json` file. The version of the `crypto` library is set to `^1.0.1`, which means any version that is `1.0.1` or higher, but less than `2.0.0`, is acceptable.